Parser classes reformatted
This is the first of a bunch of commits with the goal of refactoring the
parser classes to be human readable, so that we can modify them directly
instead of having to do so via the grammar in the .jj file.

The classes as they are *should* still work, not much functional
modification has been done, but some has, so things may be broken.