Tons of manual static analysis on the parser files
The files are now guaranteed to have the following qualities:
- The only things that are static are things that are (1) initialized in
  place, and (2) not modified at any time. Everything else is an instance
  member.
- Every public method or field needs to be public. Everything else is
  private.
- Every field that is not final needs to be mutable. Everything else is
  final.
- Everything is used. Everything unused was removed.
- All fields exist at the top of each class, followed by constructors,
  then public methods, then private methods.

Next stage is refactoring and automated analysis using PDM and FindBugs.
Refactoring will be renaming things to be more descriptive, and
restructuring the code to be more human readable.
Automated analysis will be used to scope out items that can be simplified.